Hoist RollDie/RollDice into shared src/utils/Dice.tsx
Both character pages need generic dice rolling; extracting it out of
gloomstalker's AttackSheetStateFunctions lets the Paladin refactor
reuse it instead of duplicating.
